misc(code): move fwprovider files around
