+++ fwprovider/cluster/ha/haresource_model.go
