+++ fwprovider/cluster/options/resource_options.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package fwprovider
+package options
 
 import (
 	"context"
