+++ fwprovider/nodes/hardwaremapping/datasource_usb.go
