+++ fwprovider/nodes/hardwaremapping/errors.go
