+++ fwprovider/nodes/hardwaremapping/models.go
