+++ fwprovider/nodes/hardwaremapping/resource_hardware_mapping_test.go
