+++ fwprovider/nodes/hardwaremapping/resource_pci.go
