+++ fwprovider/nodes/hardwaremapping/resource_usb.go
