+++ fwprovider/nodes/hardwaremapping/shared.go
