+++ fwprovider/nodes/resource_download_file.go
@@ -4,7 +4,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package fwprovider
+package nodes
 
 import (
 	"context"
