+++ fwprovider/nodes/vm/cdrom/model.go
