+++ fwprovider/nodes/vm/cdrom/resource.go
