+++ fwprovider/nodes/vm/cdrom/resource_schema.go
