+++ fwprovider/nodes/vm/model.go
@@ -15,11 +15,11 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/cdrom"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/cpu"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/rng"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/vga"
 	"github.com/bpg/terraform-provider-proxmox/fwprovider/types/stringset"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/cdrom"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/cpu"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/rng"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/vga"
 	"github.com/bpg/terraform-provider-proxmox/proxmox"
 	"github.com/bpg/terraform-provider-proxmox/proxmox/api"
 )
