+++ fwprovider/nodes/vm/resource.go
@@ -22,10 +22,10 @@ import (
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 
 	"github.com/bpg/terraform-provider-proxmox/fwprovider/config"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/cdrom"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/cpu"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/rng"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/vga"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/cdrom"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/cpu"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/rng"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/vga"
 	"github.com/bpg/terraform-provider-proxmox/proxmox"
 	"github.com/bpg/terraform-provider-proxmox/proxmox/api"
 	"github.com/bpg/terraform-provider-proxmox/proxmox/cluster"
