+++ fwprovider/nodes/vm/resource_schema.go
@@ -22,11 +22,11 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/cdrom"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/cpu"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/rng"
+	"github.com/bpg/terraform-provider-proxmox/fwprovider/nodes/vm/vga"
 	"github.com/bpg/terraform-provider-proxmox/fwprovider/types/stringset"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/cdrom"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/cpu"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/rng"
-	"github.com/bpg/terraform-provider-proxmox/fwprovider/vm/vga"
 )
 
 // Schema defines the schema for the resource.
