+++ fwprovider/nodes/vm/rng/resource.go
