+++ fwprovider/nodes/vm/vga/model.go
