+++ fwprovider/nodes/vm/vga/resource.go
