+++ fwprovider/nodes/vm/vga/resource_schema.go
