+++ fwprovider/nodes/vm/vga/resource_test.go
